+++ src/postgres-result.cpp
@@ -31,6 +31,209 @@
 namespace db {
   namespace postgres {
 
+    /**************************************************************************
+     * Handle to the native PostgreSQL result.
+     **************************************************************************/
+    class ResultHandle : public std::enable_shared_from_this<ResultHandle> {
+    public:
+      PGresult                   *pgresult = nullptr;              /** libpq result        */
+      ExecStatusType              status   = PGRES_EMPTY_QUERY;    /** Current status      */
+      int                         rowNum;                          /** Current row number. */
+      Connection                 &conn;
+      std::exception_ptr          lastException;
+
+      std::function<void (uint64_t)>           done = nullptr;
+      std::function<void (std::exception_ptr)> error = nullptr;
+
+      ResultHandle(Connection &conn)
+      : conn(conn) {
+      }
+
+      ~ResultHandle() {
+        if (pgresult) {
+          PQclear(pgresult);
+        }
+      }
+
+      template <class E>
+      void throwException(std::string what) {
+        if (conn.async_) {
+          lastException = std::make_exception_ptr(E(what));
+          if (error) {
+            error(lastException);
+          }
+        }
+        else {
+          throw E(what);
+        }
+      }
+
+      void throwException(std::exception_ptr eptr) {
+        assert(conn.async_);
+        lastException = eptr;
+        if (error) {
+          error(lastException);
+        }
+      }
+
+      void asyncGetResult(std::function<void ()> callback) {
+
+        auto self(shared_from_this());
+        conn.flush([self, callback]() {
+          self->conn.consumeInput([self, callback]() {
+            try {
+              callback();
+            }
+            catch (...) {
+              self->throwException(std::current_exception());
+            }
+          });
+        });
+
+      }
+
+      // -------------------------------------------------------------------------
+      // Get the first result from the server.
+      // -------------------------------------------------------------------------
+      void first() {
+        assert(pgresult == nullptr);
+        rowNum = 0;
+        if (!conn.async_) {
+          next();
+        }
+      }
+
+      // -------------------------------------------------------------------------
+      // Read the next result from the server.
+      // -------------------------------------------------------------------------
+      void next() {
+
+        if (pgresult) {
+          assert(status == PGRES_SINGLE_TUPLE);
+          PQclear(pgresult);
+          pgresult = nullptr;
+        }
+
+        assert(!(conn.async_ && PQisBusy(conn)));
+
+        pgresult = PQgetResult(conn);
+        assert(pgresult);
+        status = PQresultStatus(pgresult);
+        switch (status) {
+          case PGRES_SINGLE_TUPLE:
+            assert(PQntuples(pgresult) == 1);
+            rowNum++;
+            break;
+
+          case PGRES_TUPLES_OK:
+            // The SELECT statement did not return any row or after the last row,
+            // a zero-row object with status PGRES_TUPLES_OK is returned; this is
+            // the signal that no more rows are expected.
+            assert(PQntuples(pgresult) == 0);
+            break;
+
+          case PGRES_BAD_RESPONSE:
+          case PGRES_FATAL_ERROR:
+            throw ExecutionException(conn.lastError());
+            break;
+
+          case PGRES_COMMAND_OK:
+            break;
+
+          default:
+            assert(false);
+            break;
+        }
+
+      }
+
+      // -------------------------------------------------------------------------
+      // Clear the previous result of the connection
+      // -------------------------------------------------------------------------
+      void discard() {
+
+        switch (status) {
+          case PGRES_COMMAND_OK:
+            do {
+              PQclear(pgresult);
+              pgresult = PQgetResult(conn);
+              if (pgresult == nullptr) {
+                status = PGRES_EMPTY_QUERY;
+              }
+              else {
+                status = PQresultStatus(pgresult);
+                switch (status) {
+                  case PGRES_COMMAND_OK:
+                    break;
+
+                  case PGRES_BAD_RESPONSE:
+                  case PGRES_FATAL_ERROR:
+                    throwException<ExecutionException>(conn.lastError());
+                    break;
+
+                  case PGRES_SINGLE_TUPLE:
+                  case PGRES_TUPLES_OK:
+                    // It is not supported to execute a multi statement sql
+                    // without fetching the result with the result iterator.
+                    assert(true);
+                    break;
+
+                  default:
+                    assert(true);
+                    break;
+                }
+              }
+            } while (status != PGRES_EMPTY_QUERY);
+            break;
+
+          case PGRES_BAD_RESPONSE:
+          case PGRES_FATAL_ERROR:
+          case PGRES_TUPLES_OK:
+            PQclear(pgresult);
+            pgresult = PQgetResult(conn);
+            assert(pgresult == nullptr);
+            status = PGRES_EMPTY_QUERY;
+            break;
+
+          case PGRES_SINGLE_TUPLE:
+            next();
+            if (status == PGRES_SINGLE_TUPLE) {
+              // All results of the previous query have not been processed, we
+              // need to cancel it.
+              conn.cancel();
+            }
+            else if (status == PGRES_TUPLES_OK) {
+              PQclear(pgresult);
+              pgresult = PQgetResult(conn);
+              status = PGRES_EMPTY_QUERY;
+            }
+            assert(pgresult == nullptr);
+            break;
+
+          case PGRES_EMPTY_QUERY:
+            break;
+
+          default:
+            assert(true);
+        }
+
+      }
+
+      /**
+       * Check if the native PostgreSQL result is set.
+       **/
+      operator bool () const {
+        return pgresult != nullptr;
+      }
+
+      /**
+       * Cast to the native PostgreSQL result.
+       **/
+      operator const PGresult *() const {
+        return pgresult;
+      }
+    };
+
 #ifdef NDEBUG
 
     #define assert_oid(expected, actual) ((void)0)
@@ -144,14 +347,14 @@ namespace db {
     }
 
     int Row::num() const noexcept {
-      return result_.num_;
+      return result_.handle_->rowNum;
     }
 
     // -------------------------------------------------------------------------
     // Tests a column for a null value.
     // -------------------------------------------------------------------------
     bool Row::isNull(int column) const {
-      assert(result_.pgresult_ != nullptr);
+      assert(*result_.handle_ != nullptr);
       return PQgetisnull(result_, 0, column) == 1;
     }
 
@@ -159,7 +362,7 @@ namespace db {
     // Get a column name.
     // -------------------------------------------------------------------------
     const char *Row::columnName(int column) const {
-      assert(result_.pgresult_ != nullptr);
+      assert(*result_.handle_);
       const char *res = PQfname(result_, column);
       assert(res);
       return res;
@@ -197,7 +400,7 @@ namespace db {
 
     template<>
     std::string Row::as<std::string>(int column) const {
-      assert(result_.pgresult_ != nullptr);
+      assert(*result_.handle_);
       if (PQgetisnull(result_, 0, column)) {
         return std::string();
       }
@@ -211,7 +414,7 @@ namespace db {
     // -------------------------------------------------------------------------
     template<>
     char Row::as<char>(int column) const {
-      assert(result_.pgresult_ != nullptr);
+      assert(*result_.handle_);
       if (PQgetisnull(result_, 0, column)) {
         return '\0';
       }
@@ -224,7 +427,7 @@ namespace db {
     // -------------------------------------------------------------------------
     template<>
     std::vector<uint8_t> Row::as<std::vector<uint8_t>>(int column) const {
-      assert(result_.pgresult_ != nullptr);
+      assert(*result_.handle_);
       assert_oid(PQftype(result_, column), BYTEAOID);
       int length = PQgetlength(result_, 0, column);
       uint8_t *data = reinterpret_cast<uint8_t *>(PQgetvalue(result_, 0, column));
@@ -334,18 +537,41 @@ namespace db {
     // Result contructor
     // -------------------------------------------------------------------------
     Result::Result(Connection &conn)
-      : Row(*this), conn_(conn), begin_(*this), end_(*this) {
-      pgresult_ = nullptr;
-      status_ = PGRES_EMPTY_QUERY;
+      : Row(*this), begin_(*this), end_(*this) {
+      handle_ = std::make_shared<ResultHandle>(conn);
+    }
+
+    Result::Result(ResultHandle &handle)
+    : Row(*this), begin_(*this), end_(*this) {
+      handle_ = handle.shared_from_this();
     }
 
     // -------------------------------------------------------------------------
-    // Destructor
+    // Cast to the native PostgreSQL result.
     // -------------------------------------------------------------------------
-    Result::~Result() {
-      if (pgresult_) {
-        PQclear(pgresult_);
-      }
+    Result::operator const PGresult *() const {
+      return handle_->pgresult;
+    }
+
+    // -------------------------------------------------------------------------
+    // Get the first result from the server.
+    // -------------------------------------------------------------------------
+    void Result::first() {
+      handle_->first();
+    }
+
+    // -------------------------------------------------------------------------
+    // Get the next result from the server.
+    // -------------------------------------------------------------------------
+    void Result::next() {
+      handle_->next();
+    }
+
+    // -------------------------------------------------------------------------
+    // Discard the result.
+    // -------------------------------------------------------------------------
+    void Result::discard() {
+      handle_->discard();
     }
 
     // -------------------------------------------------------------------------
@@ -360,7 +586,7 @@ namespace db {
     // -------------------------------------------------------------------------
     Result::iterator Result::end() {
       // If there is no result available then end() = begin()
-      return Result::iterator(status_ == PGRES_SINGLE_TUPLE ? &end_ : &begin_);
+      return Result::iterator(handle_->status == PGRES_SINGLE_TUPLE ? &end_ : &begin_);
     }
 
     // -------------------------------------------------------------------------
@@ -368,9 +594,9 @@ namespace db {
     // -------------------------------------------------------------------------
     Result::iterator Result::iterator::operator ++() {
       ptr_->result_.next();
-      if (ptr_->result_.status_ != PGRES_SINGLE_TUPLE) {
+      if (ptr_->result_.handle_->status != PGRES_SINGLE_TUPLE) {
         // We've reached the end
-        assert(ptr_->result_.status_ == PGRES_TUPLES_OK);
+        assert(ptr_->result_.handle_->status == PGRES_TUPLES_OK);
         ptr_ = &ptr_->result_.end_;
       }
       return iterator (ptr_);
@@ -380,132 +606,86 @@ namespace db {
     // Number of rows affected by the SQL command
     // -------------------------------------------------------------------------
     uint64_t Result::count() const noexcept {
-      assert(pgresult_);
-      const char *count = PQcmdTuples(pgresult_);
+      assert(handle_);
+      const char *count = PQcmdTuples(handle_->pgresult);
       char *end;
       return std::strtoull(count, &end, 10);
     }
 
-    // -------------------------------------------------------------------------
-    // Get the first result from the server.
-    // -------------------------------------------------------------------------
-    void Result::first() {
-      assert(pgresult_ == nullptr);
-      num_ = 0;
-      next();
+    /**************************************************************************
+     * Registration handlers for the asynchronous mode.
+     *************************************************************************/
+
+    Result &Result::once(std::function<void (const Row &)> callback) {
+      auto handle = handle_;
+      handle_->asyncGetResult([callback, handle]() {
+        Result result(*handle);
+        result.next();
+        callback(result);
+        if (handle->done) {
+          handle->done(result.count());
+        }
+      });
+      return *this;
     }
 
-    // -------------------------------------------------------------------------
-    // Get the next result from the server.
-    // -------------------------------------------------------------------------
-    void Result::next() {
+    Result &Result::each(std::function<bool (const Row &)> callback) {
 
-      if (pgresult_) {
-        assert(status_ == PGRES_SINGLE_TUPLE);
-        PQclear(pgresult_);
-      }
+      auto handle = handle_;
+      handle_->asyncGetResult([callback, handle]() {
+        Result result(*handle);
+        bool abort = false;
+        do {
+          result.next();
+          if (handle->status == PGRES_TUPLES_OK) {
+            // We've reached the last record.
+            if (handle->done) {
+              handle->done(result.count());
+            }
+          }
+          else {
+            abort = !callback(result);
+          }
+        } while (!abort && handle->status != PGRES_TUPLES_OK && !PQisBusy(handle->conn));
 
-      pgresult_ = PQgetResult(conn_);
-      assert(pgresult_);
-      status_ = PQresultStatus(pgresult_);
-      switch (status_) {
-        case PGRES_SINGLE_TUPLE:
-          assert(PQntuples(pgresult_) == 1);
-          num_++;
-          break;
+        if (!abort && handle->status != PGRES_TUPLES_OK && PQisBusy(handle->conn)) {
+          // There are more records to fetch from the server.
+          result.each(callback);
+        }
 
-        case PGRES_TUPLES_OK:
-          // The SELECT statement did not return any row or after the last row,
-          // a zero-row object with status PGRES_TUPLES_OK is returned; this is
-          // the signal that no more rows are expected.
-          assert(PQntuples(pgresult_) == 0);
-          break;
+      });
 
-        case PGRES_BAD_RESPONSE:
-        case PGRES_FATAL_ERROR:
-          throw ExecutionException(conn_.lastError());
-          break;
-
-        case PGRES_COMMAND_OK:
-          break;
-
-        default:
-          assert(false);
-          break;
-      }
-
+      return *this;
     }
 
-    // -------------------------------------------------------------------------
-    // Clear the previous result of the connection
-    // -------------------------------------------------------------------------
-    void Result::clear() {
-
-      switch (status_) {
-        case PGRES_COMMAND_OK:
-          do {
-            PQclear(pgresult_);
-            pgresult_ = PQgetResult(conn_);
-            if (pgresult_ == nullptr) {
-              status_ = PGRES_EMPTY_QUERY;
-            }
-            else {
-              status_ = PQresultStatus(pgresult_);
-              switch (status_) {
-                case PGRES_COMMAND_OK:
-                  break;
-
-                case PGRES_BAD_RESPONSE:
-                case PGRES_FATAL_ERROR:
-                  throw ExecutionException(conn_.lastError());
-                  break;
-
-                case PGRES_SINGLE_TUPLE:
-                case PGRES_TUPLES_OK:
-                  // It is not supported to execute a multi statement sql
-                  // without fetching the result with the result iterator.
-                  assert(true);
-                  break;
-
-                default:
-                  assert(true);
-                  break;
-              }
-            }
-          } while (status_ != PGRES_EMPTY_QUERY);
-          break;
-
-        case PGRES_BAD_RESPONSE:
-        case PGRES_FATAL_ERROR:
-        case PGRES_TUPLES_OK:
-          PQclear(pgresult_);
-          pgresult_ = PQgetResult(conn_);
-          assert(pgresult_ == nullptr);
-          status_ = PGRES_EMPTY_QUERY;
-          break;
-
-        case PGRES_SINGLE_TUPLE:
-          next();
-          if (status_ == PGRES_SINGLE_TUPLE) {
-            // All results of the previous query have not been processed, we
-            // need to cancel it.
-            conn_.cancel();
-          }
-          else if (status_ == PGRES_TUPLES_OK) {
-            PQclear(pgresult_);
-            pgresult_ = PQgetResult(conn_);
-            status_ = PGRES_EMPTY_QUERY;
-          }
-          assert(pgresult_ == nullptr);
-          break;
-
-        case PGRES_EMPTY_QUERY:
-          break;
-
-        default:
-          assert(true);
+    Result &Result::done(std::function<void (uint64_t)> callback) {
+      if (handle_->status == PGRES_TUPLES_OK) {
+        callback(count());
       }
+      else if (handle_->pgresult == nullptr) {
+        // Neither once() nor each() have been registred, the call done()
+        // trigger the fetch of the query result.
+        auto handle = handle_;
+        handle_->asyncGetResult([callback, handle]() {
+          Result result(*handle);
+          result.next();
+          callback(result.count());
+        });
+      }
+      else {
+        handle_->done = callback;
+      }
+      return *this;
+    }
 
+    Result &Result::error(std::function<void (std::exception_ptr)> callback) {
+      if (handle_->lastException) {
+        callback(handle_->lastException);
+      }
+      else {
+        handle_->error = callback;
+      }
+      return *this;
     }
 
   } // namespace postgres